Feature/back member refactoring

See merge request webmobile2-sub3/s02p13a205!31

+++ back_end/A205/src/main/java/com/a205/service/MemberServiceImp.java
@@ -19,6 +19,7 @@ import com.a205.dao.MemberDAOImp;
 import com.a205.dao.MemberHasCategoryDAO;
 import com.a205.dao.MemberHasRegionDAO;
 import com.a205.dao.RegionDAO;
+import com.a205.dto.Category;
 import com.a205.dto.Member;
 import com.a205.dto.Region;
 import com.a205.dto.MemberException;
@@ -88,16 +89,12 @@ public class MemberServiceImp implements MemberService {
 			return true;
 		}
 	}
-	
-	
 	@Override
 	@Transactional
-	public boolean patchUpdate(String userId, MemberPatchRequest memberPatch) {
+	public boolean patchUpdate(String userId,MemberPatchRequest memberPatch) {
 		try {
 			Member member = dao.search(userId);
-			
 			Integer m_id = member.getM_id();
-			System.out.println("m_id ==" + m_id);
 			String m_age = memberPatch.getM_age();
 			String m_bgnTm = memberPatch.getM_bgnTm();
 			String m_endTm = memberPatch.getM_endTm();
@@ -107,60 +104,51 @@ public class MemberServiceImp implements MemberService {
 			if (!perfer_region.isEmpty()) {
 				// 기존 선호 지역 삭제
 				List<Member_has_region>list_of_member_has_regions = member_has_region_dao.searchByM_id(m_id);
-				System.out.println(list_of_member_has_regions);
-				for(Member_has_region region_obj:list_of_member_has_regions) {
-					System.out.println(region_obj);
-					boolean member_has_region_deleted = member_has_region_dao.remove(region_obj);
-					System.out.println("지워졌니?");
-					System.out.println(member_has_region_deleted);
+				if (list_of_member_has_regions.size() > 0) {
+					System.out.println(list_of_member_has_regions);
+					for(Member_has_region region_obj:list_of_member_has_regions) {
+						System.out.println(region_obj);
+						boolean member_has_region_deleted = member_has_region_dao.remove(region_obj);
+						System.out.println("지워졌니?");
+						System.out.println(member_has_region_deleted);
+					}
 				}
-				
-				
 				List<String> list = java.util.Arrays.asList(perfer_region.split(" "));
-				for(String sido:list) {
-					List<String> region_bucket = java.util.Arrays.asList(sido.split("-"));
+				for(String r_id:list) {
 					Member_has_region  member_has_region = new Member_has_region();
 					member_has_region.setM_id(m_id);
-					// r_sidoCd
-					System.out.println(region_bucket.get(0));
-					
-					// r_gugunCd
-					System.out.println(region_bucket.get(1));
-					int r_id = regionDao.selectOne(region_bucket.get(0), region_bucket.get(1)).getR_id();
-					member_has_region.setR_id(r_id);
+					Integer r_id1 = Integer.parseInt(r_id);
+					member_has_region.setR_id(r_id1);
 					boolean member_has_region_created = member_has_region_dao.add(member_has_region);
 					System.out.println(member_has_region_created);
 				}		
-			}
-			
+			}	
 			// 선호 봉사 정보 수정
 			String perfer_category = memberPatch.getPrefer_category();
 			// 만약 새로 받은 정보가 있다면 수정, 아니면 그대로 냅둬라
 			if (!perfer_category.isEmpty()) {
 				// 기존 선호 카테고리 삭제
 				List<Member_has_category>list_of_member_has_categories = member_has_category_dao.searchByM_id(m_id);
-				System.out.println(list_of_member_has_categories);
-				for(Member_has_category category_obj:list_of_member_has_categories) {
-					member_has_category_dao.remove(category_obj);
-				}
+				if (list_of_member_has_categories.size() > 0) {
 
+					System.out.println(list_of_member_has_categories);
+					for(Member_has_category category_obj:list_of_member_has_categories) {
+						member_has_category_dao.remove(category_obj);
+					}
+				}
 				// 새로운 선호 카테고리 생성
 				List<String> list2 = java.util.Arrays.asList(perfer_category.split(" "));
-				for(String cate:list2) {
-					List<String> cate_bucket = java.util.Arrays.asList(cate.split("-"));
+				for(String ca_highCd:list2) {
 					Member_has_category member_has_category = new Member_has_category();
 					member_has_category.setM_id(m_id);
-					// r_sidoCd
-					System.out.println(cate_bucket.get(0));
-					
-					// r_gugunCd
-					System.out.println(cate_bucket.get(1));
-					Integer ca_id = categoryDao.selectOne(cate_bucket.get(0),cate_bucket.get(1)).getCa_id();
-					member_has_category.setCa_id(ca_id);
-					boolean member_has_category_created = member_has_category_dao.add(member_has_category);
-					System.out.println(member_has_category_created);
+					List<Category> categories_selected_by_highCd = categoryDao.selectListByHigiCd(ca_highCd);			
+					for(Category cate:categories_selected_by_highCd) {
+						Integer ca_id = cate.getCa_id();
+						member_has_category.setCa_id(ca_id);
+						boolean member_has_category_created = member_has_category_dao.add(member_has_category);
+						System.out.println(member_has_category_created);
+					}
 				}		
-
 			}
 			member.setM_age(m_age); 
 			member.setM_bgnTm(m_bgnTm);
@@ -178,14 +166,30 @@ public class MemberServiceImp implements MemberService {
 		return false;
 	}
 
+
+	@Override
+	public boolean update(Member member) {
+		return dao.update(member);
+	}
+
+	@Override
+	public boolean remove(String id) {
+		return dao.remove(id);
+	}
+
+	@Override
+	public boolean add(Member member) {
+		return dao.add(member);
+	}
+
 	@Override
 	@Transactional
 	public Member_detail searchDetail(String userId) {
 		
 		Member member = dao.search(userId);
 		Integer m_id = member.getM_id();
-		List<Member_has_category> m_prefer_category = member_has_category_dao.searchByM_id(m_id);
-		List<Member_has_region> m_prefer_region = member_has_region_dao.searchByM_id(m_id);
+		List<Integer> m_prefer_category = member_has_category_dao.searchByM_id(m_id);
+		List<Integer> m_prefer_region = member_has_region_dao.searchByM_id(m_id);
 		
 		
 		Member_detail member_detail = new Member_detail();
@@ -209,18 +213,4 @@ public class MemberServiceImp implements MemberService {
 		return dao.searchPost(m_id);
 	}
 
-	@Override
-	public boolean update(Member member) {
-		return dao.update(member);
-	}
-
-	@Override
-	public boolean remove(String id) {
-		return dao.remove(id);
-	}
-
-	@Override
-	public boolean add(Member member) {
-		return dao.add(member);
-	}
 }
